+++ src/Setting/Settings.php
@@ -47,6 +47,8 @@ final class Settings
     public const PUI_CUSTOMER_SERVICE_INSTRUCTIONS = self::SYSTEM_CONFIG_DOMAIN . 'puiCustomerServiceInstructions';
     public const INSTALLMENT_BANNER_ENABLED = self::SYSTEM_CONFIG_DOMAIN . 'installmentBannerEnabled';
     public const LOGGING_LEVEL = self::SYSTEM_CONFIG_DOMAIN . 'loggingLevel';
+    public const EXCLUDED_PRODUCT_IDS = self::SYSTEM_CONFIG_DOMAIN . 'excludedProductIds';
+    public const EXCLUDED_PRODUCT_STREAM_IDS = self::SYSTEM_CONFIG_DOMAIN . 'excludedProductStreamIds';
 
     /**
      * @deprecated tag:v6.0.0 - Will be removed without replacement.
@@ -95,6 +97,8 @@ final class Settings
         self::INSTALLMENT_BANNER_ENABLED => true,
         self::LOGGING_LEVEL => Logger::WARNING,
         self::PUI_CUSTOMER_SERVICE_INSTRUCTIONS => 'Details zum Kundenservice finden Sie auf unserer Webseite',
+        self::EXCLUDED_PRODUCT_IDS => [],
+        self::EXCLUDED_PRODUCT_STREAM_IDS => [],
     ];
 
     public const MERCHANT_LOCATION_GERMANY = 'germany';